Phase2 start: Location, Food, User edited. Order class has been added to shopping cart.

+++ src/main/java/ie/CAs/Location.java
@@ -1,27 +1,27 @@
 package ie.CAs;
 
 public class Location {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public Location(float x, float y) {
+    public Location(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
